requirements.txt
attrs==24.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.2
pluggy==1.5.0
pymongo==4.0.1
pytest==7.2.2
pytest-flask==1.3.0
python-dotenv==0.19.2
Werkzeug==3.1.3
